refactor: unify auto form implementation
