+++ README.md
@@ -12,10 +12,9 @@ feeding it with the output of [`z.toJSONSchema`](https://zod.dev/?id=json-schema
 directly renderable (`replaceRefs` in [`auto-form.tsx`](src/components/autoform/auto-form.tsx)).
 - **Renderer:** [`AutoField`](src/components/autoform/auto-field.tsx) picks a basic control for each property based on its `type`,
 `format`, and helpers like `enum`, `anyOf`, and `additionalProperties`.
-- **Hook-based Form:** [`HookAutoForm`](src/components/autoform/hook-auto-form.tsx) integrates with React Hook Form and, as of this
-update, visually marks required fields based on the JSON Schema `required` array (part 1 of the required/validation feature).
-- **Status:** intentionally lightweight – there is no React Hook Form integration yet, and validation/submission wiring is out of
-scope for this prototype.
+- **Form wiring:** [`AutoForm`](src/components/autoform/auto-form.tsx) integrates with React Hook Form, visually marks required
+fields based on the JSON Schema `required` array, and exposes submission/reset hooks for consumers.
+- **Status:** intentionally lightweight – validation/submission beyond the provided wiring is still evolving for this prototype.
 
 ---
 
@@ -33,9 +32,9 @@ and respect any `propertyNames.pattern` constraint for the key field.
 - `anyOf` chooses the first option for now (future work could surface all variants).
 - The literal `null` type renders as a static "null" placeholder.
 
-### Required fields (HookAutoForm)
+### Required fields
 
-When using `HookAutoForm`:
+When using `AutoForm`:
 
 - Top-level required fields are determined from the schema's `required` array and are marked in the UI with an asterisk next to the
   generated label.
@@ -96,7 +95,6 @@ This is a prototype; important gaps remain:
 
 - Only the first branch of an `anyOf` is displayed.
 - Arrays are rendered as a single set of controls (no add/remove UI yet).
-- AutoForm (non-hook) does not currently mark required fields. The required-field UI applies to HookAutoForm.
 - There is no validation for required fields yet (this is part 2). This update only covers visual indication and ARIA attributes.
 - Formats beyond the ones listed above fall back to plain inputs.
 - Complex widgets (files, discriminated unions, recursive data) need dedicated UX.
